fix setpoint type float

+++ backend/db/migration/000003_create_devices_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE devices (
     id INT AUTO_INCREMENT PRIMARY KEY,
     house_id INT NOT NULL,
     climate_data_id INT NOT NULL,
-    set_point INT,
+    set_point FLOAT,
     duration INT,
     created_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
     updated_at timestamp NOT NULL DEFAULT CURRENT_TIMESTAMP,
